Code format.

+++ test_cuda/samples/5_Simulations/particles/shaders.cpp
@@ -12,9 +12,9 @@
 #define STRINGIFY(A) #A
 
 // vertex shader
-const char *vertexShader = STRINGIFY(
-    uniform float pointRadius; // point size in world space
-    uniform float pointScale;  // scale to calculate size in pixels
+const char* vertexShader = STRINGIFY(
+    uniform float pointRadius;  // point size in world space
+    uniform float pointScale;   // scale to calculate size in pixels
     uniform float densityScale; uniform float densityOffset; void main() {
       // calculate window-space point size
       vec3 posEye = vec3(gl_ModelViewMatrix * vec4(gl_Vertex.xyz, 1.0));
@@ -28,7 +28,7 @@ const char *vertexShader = STRINGIFY(
     });
 
 // pixel shader for rendering points as shaded spheres
-const char *spherePixelShader = STRINGIFY(void main() {
+const char* spherePixelShader = STRINGIFY(void main() {
   const vec3 lightDir = vec3(0.577, 0.577, 0.577);
 
   // calculate normal from texture coordinates
@@ -36,8 +36,7 @@ const char *spherePixelShader = STRINGIFY(void main() {
   N.xy = gl_TexCoord[0].xy * vec2(2.0, -2.0) + vec2(-1.0, 1.0);
   float mag = dot(N.xy, N.xy);
 
-  if (mag > 1.0)
-    discard; // kill pixels outside circle
+  if (mag > 1.0) discard;  // kill pixels outside circle
 
   N.z = sqrt(1.0 - mag);
